requirements.txt
# Flask Web Framework Dependencies
flask==3.1.1
flask-cors==6.0.0
werkzeug>=2.3.0

# Production WSGI Server (Windows compatible)
waitress>=2.1.2

# Authentication & Security
PyJWT>=2.8.0
msal>=1.24.0

# Push Notifications
pywebpush>=1.14.0
aiohttp>=3.13.3

# Database
pyodbc>=4.0.39

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests==2.32.4

# Rate Limiting
Flask-Limiter>=3.5.0

